# schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VoxForge API",
    "description": "Text-to-speech synthesis API with voice profile management",
    "version": "1.0.0"
  },
  "paths": {
    "/api/synthesize": {
      "post": {
        "tags": [
          "synthesis"
        ],
        "summary": "Synthesize text to audio",
        "description": "Convert text to audio and return the generated file.\n\nRoutes to XTTS v2 (voice cloning) if the profile has a voice sample,\notherwise uses Edge-TTS (Microsoft neural voices).\nAutomatically cancels if the client disconnects mid-generation.",
        "operationId": "synthesize_text_api_synthesize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SynthesisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/synthesize/progress/{job_id}": {
      "get": {
        "tags": [
          "synthesis"
        ],
        "summary": "Get Progress",
        "description": "Return the current progress snapshot for a running synthesis job.",
        "operationId": "get_progress_api_synthesize_progress__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/synthesize/resume/{job_id}": {
      "post": {
        "tags": [
          "synthesis"
        ],
        "summary": "Resume a crashed synthesis job",
        "operationId": "resume_job_api_synthesize_resume__job_id__post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/synthesize/incomplete/{job_id}": {
      "delete": {
        "tags": [
          "synthesis"
        ],
        "summary": "Discard a crashed job",
        "operationId": "discard_job_api_synthesize_incomplete__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Discard Job Api Synthesize Incomplete  Job Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobProgressResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "chunks_done": {
            "type": "integer",
            "title": "Chunks Done"
          },
          "chunks_total": {
            "type": "integer",
            "title": "Chunks Total"
          },
          "current_step": {
            "type": "string",
            "title": "Current Step"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "chunks_done",
          "chunks_total",
          "current_step"
        ],
        "title": "JobProgressResponse",
        "description": "Real-time progress snapshot for a synthesis job."
      },
...
      "SynthesisRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 500000,
            "minLength": 1,
            "title": "Text"
          },
          "voice_id": {
            "type": "string",
            "title": "Voice Id",
            "description": "Voice ID (e.g. es-ES-AlvaroNeural)"
          },
          "output_format": {
            "type": "string",
            "title": "Output Format",
            "description": "mp3 | wav | ogg | flac",
            "default": "mp3"
          },
          "speed": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 50.0,
            "title": "Speed",
            "description": "Speed in % (50-200)",
            "default": 100
          },
          "pitch": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": -10.0,
            "title": "Pitch",
            "description": "Pitch in semitones",
            "default": 0
          },
          "volume": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Volume",
            "description": "Volume in %",
            "default": 80
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id",
            "description": "Profile ID (optional)"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "artist": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "album": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Album"
          },
          "track_number": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9999.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Track Number"
          }
        },
        "type": "object",
        "required": [
          "text",
          "voice_id"
        ],
        "title": "SynthesisRequest",
        "description": "Text-to-speech synthesis request."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}